added an event queue
The initial goal was to expose some sort of a global error list (that could be consumed via an API) - some way of propagating internal messages to an API client (ie. tried to load this check, but couldn't because of XYZ). This has since evolved a bit further and is now instead a generic event queue. You can send whatever message to it and it'll end up in etcd.

- individual components can now send arbitrary messages to the event queue and those messages will be posted to etcd
    - usage `T.EQClient.Add(key, value)`, where key is an identifier string ("history", "error", "api", etc.) and value is an associated message