GetCheckKeysWithTags && GetClusterMembersWithTags
- added two new dal methods for fetching check keys and cluster members with tags
- still need to update director to make use of them
- still need to update `performCheckDistribution()` to use the new data

TODO: A failure of `parseTags` shouldn't be a "critical" error - the error *should* probably be trickled up to the event queue though. In other words, if a particular check does not appear to have "correct" `tags` syntax, we should heavily *log* the error and skip it, so other checks continue getting processed.