Moving back to simpler/hackier way to test log.Fatal
So looks like my original plan to use a fake logger (FieldLogger) won't work due to NewEntry() accepting only a *log.Logger and NOT an interface. So, thanks to @talpert's suggestion, moving forward with replacing the `Out` with a bufio writer and testing cases by reading that instead.

It's a little unfortunate as this was a breeze when doing with `log15`...